Test: migrate checkspec
